updated provisioning

+++ vagrant/CentOS/scripts/k8s-master.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
+echo "Executing ${0}"
+
+echo "================================================================"
+echo 
+echo "Configuring Kubernetes Master:"
+echo "  - cluster init"
+echo "  - Pod network: Flannel"
+echo 
+echo "================================================================"
+
 IPADDR=$1
 TOKEN=$2
 
@@ -15,8 +25,11 @@ if [ ! -e /etc/kubernetes/kubelet.conf ]; then
     mkdir -p $HOME/.kube
     /bin/cp -f /etc/kubernetes/admin.conf $HOME/.kube/config
     chown $(id -u):$(id -g) $HOME/.kube/config
-    rm -rf /vagrant/.kube
-    /bin/cp -rf $HOME/.kube /vagrant/.kube
+
+    if [ -d /vagrant ]; then
+        rm -rf /vagrant/.kube
+        /bin/cp -rf $HOME/.kube /vagrant/.kube
+    fi
 
     # Installing a Pod Network
     # kubectl apply -f https://raw.githubusercontent.com/coreos/flannel/v0.9.1/Documentation/kube-flannel.yml
